vmprobe build: cast the uint32_t fields in the vmstack stats line
depth_max and seg_live_max are uint32_t -- unsigned int on the x86-64 host,
long unsigned int on xtensa -- and were printed with a bare %u. Every host
build accepted it, and nothing built a CONFIG_POCKET_VM_PROBE firmware after
L2a introduced the line, so the probe-on firmware has failed -Werror=format
since 7e1c1af. L2a's integration built only the probe-off image; L2b's
integration built all three (flat on, flat off, probe on) and that is how it
surfaced.

Casts at the call, with the reason next to them, since this is exactly the
host-type-differs-from-firmware-type trap CLAUDE.md warns about.

Verified: probe-on firmware builds (DIRAM 121,484, probe cost +6,016 as
before); flat-on default and flat-off firmware build with DIRAM 115,468
unchanged; host corpus 45/45 on o2 and asan; budget_probe.sh, which parses
this line, still 10/10; G1 still 0.000 NOT_PROPORTIONAL on the flat default.

+++ components/quickjs-ng/quickjs-ng/quickjs-vm.c
@@ -58,13 +58,17 @@ void vmtest_vmstack_report(JSRuntime *rt, void *out)
 #else
     const int flat = 0;
 #endif
+    // The (unsigned) casts are not decoration: depth_max and seg_live_max are
+    // uint32_t, which is unsigned int on the x86-64 host but long unsigned int
+    // on xtensa, so a bare %u passed every host build and failed -Werror=format
+    // the first time a CONFIG_POCKET_VM_PROBE firmware compiled this file.
     fprintf(f, "#info vmstack flat=%d seg_size=%zu align=%d frame_hdr=%zu pushes=%llu depth_max=%u "
                "live_max=%zu frame_max=%zu seg_live_max=%u seg_mallocs=%llu seg_frees=%llu "
                "seg_reuses=%llu dedicated=%llu fallbacks=%llu resident_max~=%zu "
                "budget=%zu budget_hits=%llu seg_refused=%llu\n",
             flat, st->seg_size, JS_VM_FRAME_ALIGN, sizeof(JSVMSeg),
-            (unsigned long long)st->pushes, st->depth_max, st->live_bytes_max, st->frame_max,
-            st->seg_live_max, (unsigned long long)st->seg_mallocs,
+            (unsigned long long)st->pushes, (unsigned)st->depth_max, st->live_bytes_max, st->frame_max,
+            (unsigned)st->seg_live_max, (unsigned long long)st->seg_mallocs,
             (unsigned long long)st->seg_frees, (unsigned long long)st->seg_reuses,
             (unsigned long long)st->dedicated, (unsigned long long)st->fallbacks,
             (size_t)st->seg_live_max * (st->seg_size + overhead),
